Lots of things!
In main, a working dfs function (to an extent) still need to address all
paths to every node.

In Router, added and removed variables Tommy placed within it for
testing. Removed variables I added, should only have merge conflicts
with white space. Also, no need for Router Constructor with no
arguments, it is implicit within the instantiation of the object.